3 Problems listed below
+ Socket module's destroy function is not correct
+ Messaging should be in another layer entirely
+ client module needs a system for interfacing with messaging module
  and handling errors

Created a close method and a create_brutal to further seperate destruction
from how the socket is setup. Also removed unnecessary tests of on liner proxy
functions.

Created a messaging module consisting of a codec and types for messaging.
These way there is a standard way to represent different messages in the
system.

Created a backend module to handle messaging and error detection.